refactor: extract duplicated Responses API adapters into shared helpers
The three Responses-to-Chat conversion helpers were byte-for-byte
identical across Gemini, Groq, and Ollama. Extract them into
providers.ResponsesViaChat/StreamResponsesViaChat so each provider's
Responses methods become one-line delegations, and consolidate the
duplicated unit tests into a single test file.
