+++ internal/providers/gemini/gemini_test.go
@@ -479,130 +479,3 @@ data: [DONE]
 		t.Error("response should end with [DONE]")
 	}
 }
-
-func TestConvertResponsesRequestToChat(t *testing.T) {
-	temp := 0.7
-	maxTokens := 1024
-
-	tests := []struct {
-		name    string
-		input   *core.ResponsesRequest
-		checkFn func(*testing.T, *core.ChatRequest)
-	}{
-		{
-			name: "string input",
-			input: &core.ResponsesRequest{
-				Model: "gemini-2.0-flash",
-				Input: "Hello",
-			},
-			checkFn: func(t *testing.T, req *core.ChatRequest) {
-				if req.Model != "gemini-2.0-flash" {
-					t.Errorf("Model = %q, want %q", req.Model, "gemini-2.0-flash")
-				}
-				if len(req.Messages) != 1 {
-					t.Errorf("len(Messages) = %d, want 1", len(req.Messages))
-				}
-				if req.Messages[0].Role != "user" {
-					t.Errorf("Messages[0].Role = %q, want %q", req.Messages[0].Role, "user")
-				}
-			},
-		},
-		{
-			name: "with instructions",
-			input: &core.ResponsesRequest{
-				Model:        "gemini-2.0-flash",
-				Input:        "Hello",
-				Instructions: "Be helpful",
-			},
-			checkFn: func(t *testing.T, req *core.ChatRequest) {
-				if len(req.Messages) < 2 {
-					t.Fatalf("len(Messages) = %d, want at least 2", len(req.Messages))
-				}
-				if req.Messages[0].Role != "system" {
-					t.Errorf("Messages[0].Role = %q, want %q", req.Messages[0].Role, "system")
-				}
-				if req.Messages[0].Content != "Be helpful" {
-					t.Errorf("Messages[0].Content = %q, want %q", req.Messages[0].Content, "Be helpful")
-				}
-			},
-		},
-		{
-			name: "with parameters",
-			input: &core.ResponsesRequest{
-				Model:           "gemini-2.0-flash",
-				Input:           "Hello",
-				Temperature:     &temp,
-				MaxOutputTokens: &maxTokens,
-			},
-			checkFn: func(t *testing.T, req *core.ChatRequest) {
-				if req.Temperature == nil || *req.Temperature != 0.7 {
-					t.Errorf("Temperature = %v, want 0.7", req.Temperature)
-				}
-				if req.MaxTokens == nil || *req.MaxTokens != 1024 {
-					t.Errorf("MaxTokens = %v, want 1024", req.MaxTokens)
-				}
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			result := convertResponsesRequestToChat(tt.input)
-			tt.checkFn(t, result)
-		})
-	}
-}
-
-func TestConvertChatResponseToResponses(t *testing.T) {
-	resp := &core.ChatResponse{
-		ID:      "gemini-123",
-		Object:  "chat.completion",
-		Model:   "gemini-2.0-flash",
-		Created: 1677652288,
-		Choices: []core.Choice{
-			{
-				Index: 0,
-				Message: core.Message{
-					Role:    "assistant",
-					Content: "Hello! How can I help you today?",
-				},
-				FinishReason: "stop",
-			},
-		},
-		Usage: core.Usage{
-			PromptTokens:     10,
-			CompletionTokens: 20,
-			TotalTokens:      30,
-		},
-	}
-
-	result := convertChatResponseToResponses(resp)
-
-	if result.ID != "gemini-123" {
-		t.Errorf("ID = %q, want %q", result.ID, "gemini-123")
-	}
-	if result.Object != "response" {
-		t.Errorf("Object = %q, want %q", result.Object, "response")
-	}
-	if result.Model != "gemini-2.0-flash" {
-		t.Errorf("Model = %q, want %q", result.Model, "gemini-2.0-flash")
-	}
-	if result.Status != "completed" {
-		t.Errorf("Status = %q, want %q", result.Status, "completed")
-	}
-	if len(result.Output) != 1 {
-		t.Fatalf("len(Output) = %d, want 1", len(result.Output))
-	}
-	if result.Usage == nil {
-		t.Fatal("Usage should not be nil")
-	}
-	if result.Usage.InputTokens != 10 {
-		t.Errorf("InputTokens = %d, want 10", result.Usage.InputTokens)
-	}
-	if result.Usage.OutputTokens != 20 {
-		t.Errorf("OutputTokens = %d, want 20", result.Usage.OutputTokens)
-	}
-	if result.Usage.TotalTokens != 30 {
-		t.Errorf("TotalTokens = %d, want 30", result.Usage.TotalTokens)
-	}
-}
