+++ internal/providers/groq/groq.go
@@ -5,9 +5,6 @@ import (
 	"context"
 	"io"
 	"net/http"
-	"strings"
-
-	"github.com/google/uuid"
 
 	"gomodel/internal/core"
 	"gomodel/internal/llmclient"
@@ -112,135 +109,12 @@ func (p *Provider) ListModels(ctx context.Context) (*core.ModelsResponse, error)
 	return &resp, nil
 }
 
-// convertResponsesRequestToChat converts a ResponsesRequest to ChatRequest for Groq
-func convertResponsesRequestToChat(req *core.ResponsesRequest) *core.ChatRequest {
-	chatReq := &core.ChatRequest{
-		Model:       req.Model,
-		Messages:    make([]core.Message, 0),
-		Temperature: req.Temperature,
-		Stream:      req.Stream,
-	}
-
-	if req.MaxOutputTokens != nil {
-		chatReq.MaxTokens = req.MaxOutputTokens
-	}
-
-	// Add system instruction if provided
-	if req.Instructions != "" {
-		chatReq.Messages = append(chatReq.Messages, core.Message{
-			Role:    "system",
-			Content: req.Instructions,
-		})
-	}
-
-	// Convert input to messages
-	switch input := req.Input.(type) {
-	case string:
-		chatReq.Messages = append(chatReq.Messages, core.Message{
-			Role:    "user",
-			Content: input,
-		})
-	case []interface{}:
-		for _, item := range input {
-			if msgMap, ok := item.(map[string]interface{}); ok {
-				role, _ := msgMap["role"].(string)
-				content := extractContentFromInput(msgMap["content"])
-				if role != "" && content != "" {
-					chatReq.Messages = append(chatReq.Messages, core.Message{
-						Role:    role,
-						Content: content,
-					})
-				}
-			}
-		}
-	}
-
-	return chatReq
-}
-
-// extractContentFromInput extracts text content from responses input
-func extractContentFromInput(content interface{}) string {
-	switch c := content.(type) {
-	case string:
-		return c
-	case []interface{}:
-		// Array of content parts - extract text
-		var texts []string
-		for _, part := range c {
-			if partMap, ok := part.(map[string]interface{}); ok {
-				if text, ok := partMap["text"].(string); ok {
-					texts = append(texts, text)
-				}
-			}
-		}
-		return strings.Join(texts, " ")
-	}
-	return ""
-}
-
-// convertChatResponseToResponses converts a ChatResponse to ResponsesResponse
-func convertChatResponseToResponses(resp *core.ChatResponse) *core.ResponsesResponse {
-	content := ""
-	if len(resp.Choices) > 0 {
-		content = resp.Choices[0].Message.Content
-	}
-
-	return &core.ResponsesResponse{
-		ID:        resp.ID,
-		Object:    "response",
-		CreatedAt: resp.Created,
-		Model:     resp.Model,
-		Provider:  resp.Provider,
-		Status:    "completed",
-		Output: []core.ResponsesOutputItem{
-			{
-				ID:     "msg_" + uuid.New().String(),
-				Type:   "message",
-				Role:   "assistant",
-				Status: "completed",
-				Content: []core.ResponsesContentItem{
-					{
-						Type:        "output_text",
-						Text:        content,
-						Annotations: []string{},
-					},
-				},
-			},
-		},
-		Usage: &core.ResponsesUsage{
-			InputTokens:  resp.Usage.PromptTokens,
-			OutputTokens: resp.Usage.CompletionTokens,
-			TotalTokens:  resp.Usage.TotalTokens,
-		},
-	}
-}
-
 // Responses sends a Responses API request to Groq (converted to chat format)
 func (p *Provider) Responses(ctx context.Context, req *core.ResponsesRequest) (*core.ResponsesResponse, error) {
-	// Convert ResponsesRequest to ChatRequest
-	chatReq := convertResponsesRequestToChat(req)
-
-	// Use the existing ChatCompletion method
-	chatResp, err := p.ChatCompletion(ctx, chatReq)
-	if err != nil {
-		return nil, err
-	}
-
-	return convertChatResponseToResponses(chatResp), nil
+	return providers.ResponsesViaChat(ctx, p, req)
 }
 
 // StreamResponses returns a raw response body for streaming Responses API (caller must close)
 func (p *Provider) StreamResponses(ctx context.Context, req *core.ResponsesRequest) (io.ReadCloser, error) {
-	// Convert ResponsesRequest to ChatRequest
-	chatReq := convertResponsesRequestToChat(req)
-	chatReq.Stream = true
-
-	// Get the streaming response from chat completions
-	stream, err := p.StreamChatCompletion(ctx, chatReq)
-	if err != nil {
-		return nil, err
-	}
-
-	// Wrap the stream to convert chat completion format to Responses API format
-	return providers.NewOpenAIResponsesStreamConverter(stream, req.Model, "groq"), nil
+	return providers.StreamResponsesViaChat(ctx, p, req, "groq")
 }
